member component complete : spring_mvc project

+++ spring_MVC/src/main/java/com/spring/controller/MemberController.java
@@ -6,15 +6,16 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.sql.SQLException;
 import java.util.Map;
-import java.util.UUID;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpSession;
 
 import org.apache.commons.io.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -22,39 +23,44 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.spring.command.MemberModifyCommand;
+import com.spring.command.MemberRegistCommand;
 import com.spring.command.SearchCriteria;
+import com.spring.dto.MemberVO;
 import com.spring.service.MemberService;
+import com.spring.utils.MakeFileName;
 
 @Controller
 @RequestMapping("/member")
 public class MemberController {
-	
+
 	@Autowired
 	private MemberService memberService;
-	
+
 	@RequestMapping("/main")
-	public void main() {}
-	
+	public void main() {
+	}
+
 	@RequestMapping("/list")
 	public ModelAndView list(SearchCriteria cri, ModelAndView mnv) throws SQLException {
 		String url = "member/list";
-		
+
 		Map<String, Object> dataMap = memberService.getMemberList(cri);
 		mnv.addAllObjects(dataMap);
 		mnv.setViewName(url);
-		
+
 		return mnv;
 	}
-	
+
 	@RequestMapping(value = "/registForm", method = RequestMethod.GET)
 	public String registForm() {
 		String url = "member/regist";
 		return url;
 	}
-	
+
 	@Resource(name = "picturePath")
 	private String picturePath;
-	
+
 	@RequestMapping(value = "/picture", method = RequestMethod.POST, produces = "text/plain;charset=utf-8")
 	@ResponseBody
 	public ResponseEntity<String> picture(@RequestParam("pictureFile") MultipartFile multi, String oldPicture)
@@ -77,7 +83,6 @@ public class MemberController {
 
 		return entity;
 	}
-	
 
 	private String savePicture(String oldPicture, MultipartFile multi) throws Exception {
 		String fileName = null;
@@ -87,7 +92,7 @@ public class MemberController {
 
 			/* 파일저장폴더설정 */
 			String uploadPath = picturePath;
-			fileName = UUID.randomUUID().toString().replace("-", "") + ".jpg";
+			fileName = MakeFileName.toUUIDFileName(multi.getOriginalFilename(), "$$");
 			File storeFile = new File(uploadPath, fileName);
 
 			storeFile.mkdirs();
@@ -95,7 +100,7 @@ public class MemberController {
 			// local HDD 에 저장.
 			multi.transferTo(storeFile);
 
-			if (!oldPicture.isEmpty()) {
+			if (oldPicture !=null && !oldPicture.isEmpty()) {
 				File oldFile = new File(uploadPath, oldPicture);
 				if (oldFile.exists()) {
 					oldFile.delete();
@@ -106,7 +111,7 @@ public class MemberController {
 
 		return fileName;
 	}
-	
+
 	@RequestMapping(value = "/getPicture", produces = "text/plain;charset=utf-8")
 	@ResponseBody
 	public ResponseEntity<byte[]> getPicture(String picture) throws Exception {
@@ -126,13 +131,108 @@ public class MemberController {
 		}
 		return entity;
 	}
+
+	@RequestMapping("/idCheck")
+	@ResponseBody
+	public ResponseEntity<String> idCheck(String id) throws Exception {
+		ResponseEntity<String> entity = null;
+
+		try {
+			MemberVO member = memberService.getMember(id);
+
+			if (member != null) {
+				entity = new ResponseEntity<String>("duplicated", HttpStatus.OK);
+			} else {
+				entity = new ResponseEntity<String>("", HttpStatus.OK);
+			}
+		} catch (SQLException e) {
+			entity = new ResponseEntity<String>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+
+		return entity;
+	}
+
+	@RequestMapping(value = "/regist", method = RequestMethod.POST)
+	public String regist(MemberRegistCommand memberReq) throws SQLException, IOException {
+		String url = "member/regist_success";
+
+		MemberVO member = memberReq.toMemberVO();
+		memberService.regist(member);
+
+		return url;
+	}
+
+	@RequestMapping(value = "/detail", method = RequestMethod.GET)
+	public String detail(String id, Model model) throws SQLException {
+
+		String url = "member/detail";
+
+		MemberVO member = memberService.getMember(id);
+		model.addAttribute("member", member);
+
+		return url;
+	}
+
+	@RequestMapping(value = "/modify", method = RequestMethod.POST)
+	public String modify(MemberModifyCommand modifyReq,HttpSession session,Model model) throws Exception {
+		String url="member/modify_success";
+
+		MemberVO member = modifyReq.toParseMember();
+
+		// 신규 파일 변경 및 기존 파일 삭제
+		String fileName = savePicture(modifyReq.getOldPicture(), modifyReq.getPicture());
+		member.setPicture(fileName);
+		
+		//파일변경 없을시 기존 파일명 유지
+		if (modifyReq.getPicture().isEmpty()) {
+			member.setPicture(modifyReq.getOldPicture());
+		}
+		
+		//DB 내용 수정
+		memberService.modify(member);
+		
+		// 로그인한 사용자의 경우 수정된 정보로 session 업로드
+		MemberVO loginUser = (MemberVO) session.getAttribute("loginUser");
+		if (loginUser != null && member.getId().equals(loginUser.getId())) {
+			session.setAttribute("loginUser", member);
+		}
+		
+		model.addAttribute("member",memberService.getMember(modifyReq.getId()));
+		
+		return url;
+
+	}
+	
+	@RequestMapping(value = "/remove", method = RequestMethod.GET)
+	public String remove(String id, HttpSession session, Model model) throws SQLException {
+		String url = "member/remove_success";
+		
+		MemberVO member;
+
+		// 이미지 파일을 삭제
+		member = memberService.getMember(id);
+
+		String savePath = this.picturePath;
+		File imageFile = new File(savePath, member.getPicture());
+		if (imageFile.exists()) {
+			imageFile.delete();
+		}
+		
+		//DB삭제
+		memberService.remove(id);
+		
+		// 삭제되는 회원이 로그인 회원인경우 로그아웃 해야함.
+		MemberVO loginUser = (MemberVO) session.getAttribute("loginUser");
+		if (loginUser.getId().equals(member.getId())) {
+			session.invalidate();
+		}
+		
+		model.addAttribute("member", member);
+		
+		return url;
+	}
+
 }
-
-
-
-
-
-
